refactor(site/src/pages/AgentsPage): split file reference chip
